Refactor Register, login method

+++ ShopOnlineApi/ShopOnlineApi/Models/Order.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public int? UserId { get; set; }
         public DateOnly? OrderDate { get; set; }
-        public double? Price { get; set; }
+        public int? Price { get; set; }
         public int? ProductId { get; set; }
 
         public virtual Product? Product { get; set; }
